requirements.txt
bilibili-api-python
asyncio
zhipuai
requests
pygame
datetime
flask
flask_socketio
flask_cors